requirements.txt
streamlit
tensorflow
pillow
numpy
azure-storage-blob
opencv-python-headless
huggingface_hub